refactor api slice structure
