+++ src/components/editor/content/profile/container.tsx
@@ -1,8 +1,10 @@
 import { useSelector } from 'react-redux';
 
-import { selectors } from '@/state/api';
+import { apiState } from '@/state/api';
 
 import { Profile } from './component';
+
+const selectors = apiState.profile.selectors;
 
 /**
  * This is the Profile component but connected to the store
