+++ src/components/editor/content/profile/controls/input-description.tsx
@@ -2,9 +2,11 @@ import * as React from 'react';
 import { useSelector } from 'react-redux';
 import styled from 'styled-components';
 
-import { selectors, useProfileUpdater } from '@/state/api';
+import { apiState } from '@/state/api';
 import { Textarea } from '@/components/editor/common/form';
 import { Profile } from '@/shared/types';
+
+const { selectors, useProfileUpdater } = apiState.profile;
 
 export function InputDescriptionContainer() {
   const description = useSelector(selectors.selectProfileDescription);
