+++ src/state/api/slice.ts
@@ -47,32 +47,7 @@ export const apiSlice = createApi({
 
 export const { useGetResumeQuery } = apiSlice;
 
-const selectResumeResult = apiSlice.endpoints.getResume.select();
-
-const selectResumeStatus = createSelector(
-  selectResumeResult,
-  (result) => result.status,
-);
-
-const selectProfile = createSelector(
-  selectResumeResult,
-  (result) => result.data?.content.profile ?? Profile.create(),
-);
-
-const selectProfileTitle = createSelector(
-  selectProfile,
-  (profile) => profile.title,
-);
-
-const selectProfileDescription = createSelector(
-  selectProfile,
-  (profile) => profile.description,
-);
-
-const selectProfileCardBackground = createSelector(
-  selectProfile,
-  (profile) => profile.cardStyle.background,
-);
+export const selectResumeResult = apiSlice.endpoints.getResume.select();
 
 export function useUpdateResume() {
   const [updateResume_, meta] = apiSlice.useUpdateResumeMutation({ fixedCacheKey: 'update-resume' });
@@ -92,45 +67,5 @@ export function useUpdateResume() {
 
   return [updateResume, meta] as const;
 }
-
-export function useProfileUpdater() {
-  const profile = useSelector(selectProfile);
-  const [updateResume] = useUpdateResume();
-
-  function updateProfile(newProfile: Partial<Profile>) {
-    updateResume({
-      profile: Profile.update(profile, newProfile),
-    });
-  }
-
-  function updateProfleTitle(title: Partial<Profile['title']>) {
-    updateProfile(Profile.updateTitle(profile, title));
-  }
-
-  function updateProfleDescription(description: Partial<Profile['description']>) {
-    updateProfile(Profile.updateDescription(profile, description));
-  }
-
-  function updateCardStyle(style: Partial<Profile['cardStyle']>) {
-    updateProfile(Profile.updateCardStyle(profile, style));
-  }
-
-  const updater = {
-    updateProfile,
-    updateProfleTitle,
-    updateProfleDescription,
-    updateCardStyle,
-  };
-
-  return [updater] as const;
-}
-
-export const selectors = {
-  selectResumeStatus,
-  selectProfile,
-  selectProfileTitle,
-  selectProfileDescription,
-  selectProfileCardBackground,
-};
 
 export default apiSlice;
